Remove tests, change attributes names and fix repositories

+++ src/main/java/cat/udl/eps/softarch/demo/repository/DonationRepository.java
@@ -3,5 +3,5 @@ package cat.udl.eps.softarch.demo.repository;
 import cat.udl.eps.softarch.demo.domain.Donation;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface DonationRepository extends PagingAndSortingRepository<Donation, Integer> {
+public interface DonationRepository extends PagingAndSortingRepository<Donation, Long> {
 }
